+++ google2/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file documents changes to the `workbench-google2` library, including notes on how to upgrade to new versions.
 
+## 0.14
+- Changes the return types for some methods in `GKEInterpreter` from `F[Operation]` to `F[Option[Operation]]`
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.14-TRAVIS-REPLACE-ME"`
+
 ## 0.13
 Changed:
 - `GKEService.createCluster` now uses legacy `com.google.api.services.container` client and model objects